GEODE-8284: Break up StringsIntegration test class (#5277)
